move stream methods to io folder
