+++ js/src/Arrow.dom.ts
@@ -18,9 +18,9 @@
 import streamAdapters from './io/adapters';
 import { RecordBatchReader } from './ipc/reader';
 import { RecordBatchWriter } from './ipc/writer';
-import { toDOMStream } from './ipc/whatwg/iterable';
-import { recordBatchReaderThroughDOMStream } from './ipc/whatwg/reader';
-import { recordBatchWriterThroughDOMStream } from './ipc/whatwg/writer';
+import { toDOMStream } from './io/whatwg/iterable';
+import { recordBatchReaderThroughDOMStream } from './io/whatwg/reader';
+import { recordBatchWriterThroughDOMStream } from './io/whatwg/writer';
 
 streamAdapters.toDOMStream = toDOMStream;
 RecordBatchReader['throughDOM'] = recordBatchReaderThroughDOMStream;
@@ -84,16 +84,22 @@ export {
     predicate,
     util,
     DataBuilder,
-    BoolBuilder,
-    IntBuilder, Int8Builder, Int16Builder, Int32Builder, Int64Builder, Uint8Builder, Uint16Builder, Uint32Builder, Uint64Builder,
-    FloatBuilder, Float16Builder, Float32Builder, Float64Builder,
-    Utf8Builder,
     BinaryBuilder,
-    FixedSizeBinaryBuilder,
+    BoolBuilder,
     DateBuilder, DateDayBuilder, DateMillisecondBuilder,
-    TimestampBuilder, TimestampSecondBuilder, TimestampMillisecondBuilder, TimestampMicrosecondBuilder, TimestampNanosecondBuilder,
-    TimeBuilder, TimeSecondBuilder, TimeMillisecondBuilder, TimeMicrosecondBuilder, TimeNanosecondBuilder,
     DecimalBuilder,
     DictionaryBuilder,
-    IntervalBuilder, IntervalDayTimeBuilder, IntervalYearMonthBuilder
+    FixedSizeBinaryBuilder,
+    FixedSizeListBuilder,
+    FloatBuilder, Float16Builder, Float32Builder, Float64Builder,
+    IntervalBuilder, IntervalDayTimeBuilder, IntervalYearMonthBuilder,
+    IntBuilder, Int8Builder, Int16Builder, Int32Builder, Int64Builder, Uint8Builder, Uint16Builder, Uint32Builder, Uint64Builder,
+    ListBuilder,
+    MapBuilder,
+    NullBuilder,
+    StructBuilder,
+    TimestampBuilder, TimestampSecondBuilder, TimestampMillisecondBuilder, TimestampMicrosecondBuilder, TimestampNanosecondBuilder,
+    TimeBuilder, TimeSecondBuilder, TimeMillisecondBuilder, TimeMicrosecondBuilder, TimeNanosecondBuilder,
+    UnionBuilder, DenseUnionBuilder, SparseUnionBuilder,
+    Utf8Builder,
 } from './Arrow';
