+++ js/src/Arrow.node.ts
@@ -18,9 +18,9 @@
 import streamAdapters from './io/adapters';
 import { RecordBatchReader } from './ipc/reader';
 import { RecordBatchWriter } from './ipc/writer';
-import { toNodeStream } from './ipc/node/iterable';
-import { recordBatchReaderThroughNodeStream } from './ipc/node/reader';
-import { recordBatchWriterThroughNodeStream } from './ipc/node/writer';
+import { toNodeStream } from './io/node/iterable';
+import { recordBatchReaderThroughNodeStream } from './io/node/reader';
+import { recordBatchWriterThroughNodeStream } from './io/node/writer';
 
 streamAdapters.toNodeStream = toNodeStream;
 RecordBatchReader['throughNode'] = recordBatchReaderThroughNodeStream;
