+++ js/src/io/node/iterable.ts
