+++ js/src/io/node/reader.ts
