+++ js/src/io/node/writer.ts
