+++ js/src/io/whatwg/iterable.ts
