+++ js/src/io/whatwg/writer.ts
